Added ricochet and flying parts to hits.

+++ core/src/main/java/com/gadarts/returnfire/components/AiComponent.kt
@@ -7,13 +7,13 @@ import com.badlogic.gdx.math.Vector3
 class AiComponent : Component {
     var returnToBase: Boolean = false
         private set
-    var lastHpCheck: Int = 0
+    var lastHpCheck: Float = 0F
     val runAway = Vector3()
     var target: Entity? = null
     var attackReadyTime: Long = 0
     var attackReady: Boolean = true
 
-    fun init(target: Entity?, initialHp: Int) {
+    fun init(target: Entity?, initialHp: Float) {
         this.target = target
         this.lastHpCheck = initialHp
         this.attackReadyTime = 0
