+++ core/src/main/java/com/gadarts/returnfire/model/definitions/TurretCharacterDefinition.kt
@@ -5,14 +5,14 @@ import com.gadarts.returnfire.assets.definitions.ModelDefinition
 import com.gadarts.returnfire.model.CharacterType
 
 enum class TurretCharacterDefinition(
-    private val hp: Int,
+    private val hp: Float,
     private val baseModelDefinition: ModelDefinition,
     private val smokeEmissionRelativePosition: Vector3,
     private val gravity: Vector3,
     private val linearFactor: Vector3,
 ) : CharacterDefinition {
-    TURRET_CANNON(40, ModelDefinition.TURRET_BASE, Vector3(0F, 2F, 0F), Vector3.Zero, Vector3.Zero),
-    TANK(100, ModelDefinition.TANK_BODY, Vector3.Zero, Vector3(0F, -10F, 0F), Vector3(1F, 1F, 1F));
+    TURRET_CANNON(40F, ModelDefinition.TURRET_BASE, Vector3(0F, 2F, 0F), Vector3.Zero, Vector3.Zero),
+    TANK(100F, ModelDefinition.TANK_BODY, Vector3.Zero, Vector3(0F, -10F, 0F), Vector3(1F, 1F, 1F));
 
     override fun isFlyer(): Boolean {
         return false
@@ -38,7 +38,7 @@ enum class TurretCharacterDefinition(
         return smokeEmissionRelativePosition
     }
 
-    override fun getHP(): Int {
+    override fun getHP(): Float {
         return hp
     }
