+++ core/src/main/java/com/gadarts/returnfire/systems/EntityBuilder.kt
@@ -95,8 +95,8 @@ interface EntityBuilder {
     ): EntityBuilder
 
     fun addAmbComponent(rotation: Float, def: AmbDefinition, scale: Vector3): EntityBuilder
-    fun addAiComponent(initialHp: Int, target: Entity? = null): EntityBuilder
-    fun addAiComponentToEntity(entity: Entity, initialHp: Int, target: Entity? = null): AiComponent
+    fun addAiComponent(initialHp: Float, target: Entity? = null): EntityBuilder
+    fun addAiComponentToEntity(entity: Entity, initialHp: Float, target: Entity? = null): AiComponent
     fun addPhysicsComponentToEntity(
         entity: Entity,
         shape: btCollisionShape,
@@ -112,7 +112,7 @@ interface EntityBuilder {
         explosion: ParticleEffectDefinition?,
         explosive: Boolean,
         friendly: Boolean,
-        damage: Int
+        damage: Float
     ): EntityBuilder
 
     fun addPhysicsComponentPooled(
